requirements.txt
# MQTT client for sensor simulator
paho-mqtt

# Kafka client for streaming data to the dashboard
kafka-python

# Streamlit dashboard
streamlit

# FastAPI server for model predictions
fastapi
uvicorn

# Deep Learning framework
tensorflow

# Data processing
pandas
numpy
scikit-learn

# Optional: for plots and data visualization
matplotlib
seaborn
